Move most of drawend functionality to OL side
